requirements.txt
numpy==1.23.0
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
Pillow==9.5.0
Pillow==10.0.1
streamlit==1.22.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
